// src/app/pages/operations/forms/vehicle/vehicle-form/vehicle-form.type.ts
//https://netbasal.com/typed-reactive-forms-in-angular-no-longer-a-type-dream-bf6982b0af28
export interface IVehicleForm {
  department: string;
  vehicleMake: string;
  year: string;
  vin: string;
  exp: string;
  vehicleNumber: string;
  mileage: string;
  lastServiceDate: string;
  typeOfService: string;
  fuelType: string;
  createdBy: number;
  active: number;
  licensePlate: string;
  createdDate: string;
}